Correciones de UserDetailsService
