+++ src/main/java/com/sintergica/apiv2/servicios/GroupService.java
@@ -8,28 +8,27 @@ import com.sintergica.apiv2.repositorio.CompanyRepository;
 import com.sintergica.apiv2.repositorio.GrantRepository;
 import com.sintergica.apiv2.repositorio.GroupRepository;
 import com.sintergica.apiv2.repositorio.UserRepository;
-import java.util.*;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 import java.util.stream.Collectors;
+
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
+@RequiredArgsConstructor
 public class GroupService {
 
   private final GroupRepository groupRepository;
   private final UserRepository userRepository;
   private final CompanyRepository companyRepository;
   private final GrantRepository grantRepository;
-
-  public GroupService(
-      GroupRepository groupRepository,
-      UserRepository userRepository,
-      CompanyRepository companyRepository,
-      GrantRepository grantRepository) {
-    this.groupRepository = groupRepository;
-    this.userRepository = userRepository;
-    this.companyRepository = companyRepository;
-    this.grantRepository = grantRepository;
-  }
 
   public List<Group> findAll() {
     return groupRepository.findAll();
@@ -54,10 +53,10 @@ public class GroupService {
     User currentUser = userRepository.findByEmail(emailUser);
 
     if (currentGroupOpt.isPresent()
-        && currentUser != null
-        && currentUser.getCompany() != null
-        && currentGroupOpt.get().getCompany() != null
-        && currentUser.getCompany().getId().equals(currentGroupOpt.get().getCompany().getId())) {
+            && currentUser != null
+            && currentUser.getCompany() != null
+            && currentGroupOpt.get().getCompany() != null
+            && currentUser.getCompany().getId().equals(currentGroupOpt.get().getCompany().getId())) {
 
       Group currentGroup = currentGroupOpt.get();
       currentGroup.getUser().add(currentUser);
@@ -70,12 +69,12 @@ public class GroupService {
 
     HashMap<String, String> response = new HashMap<>();
     response.put(
-        "error", "Uno de los campos es nulo o la empresa no esta asociada con el usuario y grupo");
+            "error", "Uno de los campos es nulo o la empresa no esta asociada con el usuario y grupo");
     return response;
   }
 
   public Map<String, String> addNewGroup(
-      String nameGroup, List<String> grantList, UUID companyUUID) {
+          String nameGroup, List<String> grantList, UUID companyUUID) {
     HashMap<String, String> response = new HashMap<>();
     Optional<Company> company = companyRepository.findById(companyUUID);
 
@@ -85,10 +84,10 @@ public class GroupService {
     }
 
     Set<Grant> grants =
-        grantList.stream()
-            .map(grantRepository::findByName)
-            .filter(Objects::nonNull)
-            .collect(Collectors.toSet());
+            grantList.stream()
+                    .map(grantRepository::findByName)
+                    .filter(Objects::nonNull)
+                    .collect(Collectors.toSet());
 
     Group newGroup = new Group();
     newGroup.setName(nameGroup);
@@ -154,4 +153,13 @@ public class GroupService {
 
     return response;
   }
+
+  public Optional<Group> findGroupById(UUID uuidGroup) {
+    return groupRepository.findById(uuidGroup);
+  }
+
+  public List<Group> findGroupsByUserIdsIn(List<UUID> userIds) {
+    return this.groupRepository.findGroupsByUserIdsIn(userIds);
+  }
+
 }
