+++ src/main/java/com/sintergica/apiv2/servicios/RolService.java
@@ -26,4 +26,7 @@ public class RolService {
         .findFirst()
         .orElse(null);
   }
+
+
+
 }
